Added new Pages and Test classes.
