+++ src/test/resources/TestNG.xml
@@ -3,11 +3,12 @@
 
 <suite name="TestNG">
 
-    <test name="Example Test" preserve-order="true">
+    <test name="Example Test" preserve-order="false">
         <parameter name="baseURL" value="https://qa.koel.app/"/>
         <classes>
             <class name="BaseTest"/>
             <class name="HomeTests"/>
+            <class name="SongsTests"/>
 
         </classes>
     </test>
